code for converting JSON to JSON-LD
json_to_json-ld.pl is a script for adding appropriate $schema and @type lines to a DATS json file to generate an equivalent JSON-LD file. dataset_reference.txt is a lookup table for identifiers and URLs used by this script.  

NB:  This covers most but not all necessary cases, because there can be ambiguous variations (such as the "authors" entity sometimes being a Person (or list thereof) and sometimes an Organisation (eg a consortium)) between input DATS.json files.